Moved same function to a trait
