+++ Block/Document.php
@@ -8,7 +8,7 @@ use Elgentos\PrismicIO\ViewModel\DocumentResolver;
 use Elgentos\PrismicIO\ViewModel\LinkResolver;
 use Magento\Framework\View\Element\Context;
 
-class Document extends AbstractBlock implements BlockInterface
+class Document extends AbstractBlock
 {
 
     /** @var Api */
@@ -71,22 +71,4 @@ class Document extends AbstractBlock implements BlockInterface
         $this->setDocument($document);
     }
 
-    /**
-     * Update child document to use relative paths
-     *
-     * @param string $alias
-     * @return bool
-     * @throws \Elgentos\PrismicIO\Exception\ContextNotFoundException
-     * @throws \Elgentos\PrismicIO\Exception\DocumentNotFoundException
-     */
-    public function updateChildDocument(string $alias): bool
-    {
-        $block = $this->getChildBlock($alias);
-        if (! $block) {
-            return false;
-        }
-
-        $block->setDocument($this->getDocument());
-        return true;
-    }
 }
